use background graphite writer
this lets us retry on failures instead of just bailing out
and also it stops the needless opening of multiple concurrent
tcp connections

in particular, this also solves a mysterious problem we were seeing in
https://github.com/raintank/metrictank/issues/277 where we were getting
i/o timeouts for writes that returned in microseconds. (though submit()
was taking a few seconds).
Didn't get to the bottom of it.